feat(app2): sorted-balances store
