+++ app2/src/routes/balances/+page.svelte
@@ -16,84 +16,7 @@ import type { Tokens } from "$lib/schema/token"
 // Example wallet address - this would come from wallet connection in real app
 const testAddress = AddressEvmCanonical.make("0xe6831e169d77a861a0e71326afa6d80bcc8bc6aa")
 
-const getSortedTokens = (
-  tokens: Tokens,
-  chain: Chain,
-  bs: BalancesStore,
-  address: AddressCanonicalBytes
-) =>
-  tokens
-    .map(token => {
-      const balance = bs.getBalance(chain.universal_chain_id, address, token.denom)
-      const error = bs.getError(chain.universal_chain_id, address, token.denom)
-      const tokenInfo = tokensStore
-        .getData(chain.universal_chain_id)
-        .pipe(
-          Option.flatMap(tokens => Option.fromNullable(tokens.find(t => t.denom === token.denom)))
-        )
-
-      // Get decimals from token info
-      const decimals =
-        Option.getOrNull(
-          Option.flatMap(tokenInfo, t => Option.fromNullable(t.representations[0]?.decimals))
-        ) ?? 18 // Default to 18 if not found
-
-      // Calculate numeric value for sorting
-      const numericValue = Option.match(balance, {
-        onNone: () => -1n,
-        onSome: bal =>
-          Option.match(Option.fromNullable(bal), {
-            onNone: () => 0n,
-            onSome: val => val
-          })
-      })
-
-      return {
-        token,
-        balance,
-        error,
-        numericValue,
-        decimals
-      }
-    })
-    .sort((a, b) => {
-      // First, separate by status
-      if (Option.isSome(a.error) && !Option.isSome(b.error)) return 1
-      if (!Option.isSome(a.error) && Option.isSome(b.error)) return -1
-
-      if (Option.isNone(a.balance) && Option.isSome(b.balance)) return 1
-      if (Option.isSome(a.balance) && Option.isNone(b.balance)) return -1
-
-      // Then sort by value
-      if (a.numericValue === -1n && b.numericValue !== -1n) return 1
-      if (a.numericValue !== -1n && b.numericValue === -1n) return -1
-
-      if (a.numericValue === 0n && b.numericValue > 0n) return 1
-      if (a.numericValue > 0n && b.numericValue === 0n) return -1
-
-      // Sort by actual value if both have balances
-      if (a.numericValue > 0n && b.numericValue > 0n) {
-        // Adjust for decimals
-        const aAdjusted = a.numericValue / 10n ** BigInt(a.decimals)
-        const bAdjusted = b.numericValue / 10n ** BigInt(b.decimals)
-        return aAdjusted < bAdjusted ? 1 : -1
-      }
-
-      return 0
-    })
-
-const sortedBalances = $derived(
-  chains.data.pipe(
-    Option.map(d =>
-      d.map(chain => ({
-        chain,
-        tokens: tokensStore
-          .getData(chain.universal_chain_id)
-          .pipe(Option.map(ts => getSortedTokens(ts, chain, balancesStore, testAddress)))
-      }))
-    )
-  )
-)
+import { sortedBalancesStore } from "$lib/stores/sorted-balances.svelte"
 
 function fetchAllBalances() {
   const chainsData = Option.getOrNull(chains.data)
@@ -151,12 +74,12 @@ $effect(() => {
       {#if chain.rpc_type !== "aptos"}
         <div class="flex flex-col">
           
-          {#if Option.isNone(sortedBalances)}
+          {#if Option.isNone(sortedBalancesStore.sortedBalances)}
             <Card>
               <div class="text-zinc-500">Loading balances...</div>
             </Card>
           {:else}
-            {@const tokensForChain = Option.fromNullable(sortedBalances.value.find(v => v.chain.universal_chain_id === chain.universal_chain_id)).pipe(Option.flatMap(c => c.tokens))}
+            {@const tokensForChain = Option.fromNullable(sortedBalancesStore.sortedBalances.value.find(v => v.chain.universal_chain_id === chain.universal_chain_id)).pipe(Option.flatMap(c => c.tokens))}
             {#if Option.isNone(tokensForChain)}
               <Card>
                 <div class="text-zinc-500">No balances found</div>
